// separate_audio_streams_per_participant_download_audio_files/src/schemas/AudioSeparateRawDataSchema.ts
import { z } from "zod";

/**
 * Schema for the audio_separate_raw.data event
 */
export const AudioSeparateRawDataSchema = z.object({
    "event": z.literal("audio_separate_raw.data"),
    "data": z.object({
        "data": z.object({
            "buffer": z.string(), // base64-encoded raw audio 16 kHz mono, S16LE(16-bit PCM LE)
            "timestamp": z.object({ // Timestamp of the first byte in the buffer. More info about timestamps: https://docs.recall.ai/docs/download-schemas#/schema-timestamps
                "relative": z.number(), // "Timestamp in seconds"),
                "absolute": z.string(), // "ISO 8601 absolute timestamp (e.g. 2025-01-01 00:00:00)")
            }),
            "participant": z.object({
                "id": z.number(), // Recall.ai assigned participant id (e.g. 100, 200, 300)
                "name": z.string().nullable(), // Display name from meeting
                "is_host": z.boolean(), // True if the participant is the host
                "platform": z.string().nullable(), // Meeting platform constant. values: 'desktop', 'dial-in', 'unknown'
                "extra_data": z.any(), // Extra data about the participant from the meeting platform
                "email": z.string().nullish(), // Email address of the participant if using Recall's calendar integration
            }),
        }),
        "realtime_endpoint": z.object({
            "id": z.string(),
            "metadata": z.record(z.string(), z.string()),
        }),
        "audio_separate": z.object({
            "id": z.string(),
            "metadata": z.record(z.string(), z.string()),
        }),
        "recording": z.object({
            "id": z.string(),
            "metadata": z.record(z.string(), z.string()),
        }),
        "bot": z.object({
            "id": z.string(),
            "metadata": z.record(z.string(), z.string()),
        }).nullish(),
    }),
});